Fix i18n checker
