--- tools/i18n_checker/src/hclua
@@ -1,1 +1,0 @@
-../../../runtime/data/script/kernel/hclua
