+++ tools/i18n_checker/src/hclua/decoder.lua
@@ -1,4 +1,4 @@
-local Util = require_relative("hclua.util")
+local Util = require("hclua.util")
 
 local Decoder = {}
