+++ tools/i18n_checker/src/hclua/lexer.lua
